requirements/base.txt
Django==1.7.1
Pillow==2.7.0
PyJWT==0.4.0
Unipath==1.0
django-braces==1.4.0
django-crispy-forms==1.4.0
django-filebrowser==3.5.7
django-grappelli==2.5.6
django-model-utils==2.2
httplib2==0.9
oauth2==1.5.211
oauthlib==0.7.2
py==1.4.26
python-openid==2.2.5
python-social-auth==0.2.1
requests==2.5.1
requests-oauthlib==0.4.2
six==1.8.0
sqlparse==0.1.14
tox==1.8.1
virtualenv==12.0.5
wsgiref==0.1.2
